// src/services/generated/apiTypes/maritalStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Laravel
 * OpenAPI spec version: 0.0.1
 */

export type MaritalStatus = typeof MaritalStatus[keyof typeof MaritalStatus];


export const MaritalStatus = {
  single: 'single',
  married: 'married',
  divorced: 'divorced',
  widowed: 'widowed',
  cohabiting: 'cohabiting',
} as const;